refactor: reorganize application services and domain structure for improved readability
