+++ Orderly.UsersService/src/Orderly.Users.Application/Models/RegisterRequest.cs
@@ -1,4 +1,4 @@
-﻿using Orderly.Users.Domain;
+﻿using Orderly.Users.Domain.Users;
 
 namespace Orderly.Users.Application.Models;
