+++ Orderly.UsersService/src/Orderly.Users.Domain/Users/ValueObjects/Email.cs
@@ -1,0 +1,3 @@
+﻿namespace Orderly.Users.Domain.Users.ValueObjects;
+
+public sealed record Email(string Value);
