+++ Orderly.UsersService/src/Orderly.Users.Domain/Users/ValueObjects/Name.cs
@@ -1,0 +1,3 @@
+﻿namespace Orderly.Users.Domain.Users.ValueObjects;
+
+public sealed record Name(string Value);
