+++ Orderly.UsersService/src/Orderly.Users.Domain/Users/ValueObjects/Password.cs
@@ -1,0 +1,4 @@
+﻿namespace Orderly.Users.Domain.Users.ValueObjects;
+
+public sealed record Password(string Value);
+
