--- Orderly.UsersService/src/Orderly.Users.Domain/ValueObjects/Email.cs
@@ -1,3 +1,0 @@
-﻿namespace Orderly.Users.Domain.ValueObjects;
-
-public sealed record Email(string Value);
