--- Orderly.UsersService/src/Orderly.Users.Domain/ValueObjects/Name.cs
@@ -1,3 +1,0 @@
-﻿namespace Orderly.Users.Domain.ValueObjects;
-
-public sealed record Name(string Value);
