--- Orderly.UsersService/src/Orderly.Users.Domain/ValueObjects/Password.cs
@@ -1,4 +1,0 @@
-﻿namespace Orderly.Users.Domain.ValueObjects;
-
-public sealed record Password(string Value);
-
